LayoutTests/js/dom/modules/module-load-event.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test module execution order between tags.');
// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
function onLoad()
{
    shouldBeTrue(`moduleExecuted`);
    finishJSTest();
}
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module" onload="onLoad()">
debug('Executing an inlined module.');
window.moduleExecuted = true;
</script>
</body>
</html>
