LayoutTests/js/dom/modules/module-load-event-expected.txt

Test module execution order between tags.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Executing an inlined module.
PASS moduleExecuted is true
PASS successfullyParsed is true

TEST COMPLETE
